[FIX] Mantis 0017788: [DataCollection] Files no longer get deleted after a second save action.

+++ Modules/DataCollection/classes/class.ilDataCollectionRecordField.php
@@ -160,7 +160,7 @@ class ilDataCollectionRecordField {
 				$this->value = $tmp;
 				//delete old file from filesystem
 				// TODO Does not belong here, create separate class ilDataCollectionFileField and overwrite setValue method
-				if ($old && $this->field->getDatatypeId() == ilDataCollectionDatatype::INPUTFORMAT_FILE) {
+				if ($old && $old != $this->value && $this->field->getDatatypeId() == ilDataCollectionDatatype::INPUTFORMAT_FILE) {
 					$this->record->deleteFile($old);
 				}
 			}
